+++ R/check_micromamba_version.R
@@ -1,6 +1,7 @@
 check_micromamba_version <- function(
-    umamba_path = NULL,
-    target_version = "2.1.1") {
+  umamba_path = NULL,
+  target_version = "2.3.0"
+) {
   # rlang::is_string(target_version)
   version_string <- get_micromamba_version(umamba_path)
   compare_res <- utils::compareVersion(version_string, target_version)